add: _0_0_Install_Latest_MooaToon_Precompiled_And_Override_All_Modified_Files add: _0_1_Run_MooaToon_Precompiled_Project

+++ ReleaseTools/Release.py
@@ -7,9 +7,15 @@ import github_release as ghr
 
 ue_version = "5.1"
 repo_name = "JasonMa0012/MooaToon"
-mooatoon_root_path = r"E:\MooaToon"
+mooatoon_root_path = r"X:\MooaToon"
 if len(sys.argv) > 1:
     mooatoon_root_path = sys.argv[1]
+
+argv = [
+    '--Reupload'
+]
+if len(sys.argv) > 2:
+    argv = sys.argv[2:]
 
 bandizip_path = mooatoon_root_path + r"\InstallTools\BANDIZIP-PORTABLE\Bandizip.x64.exe"
 zip_path = mooatoon_root_path + r"\ReleaseTools\Zip"
@@ -32,48 +38,62 @@ def AsyncRun(args):
     return process.poll()
 
 
-print("======Clean======")
-for file_name in os.listdir(zip_path):
-    file_path = os.path.join(zip_path, file_name)
-    os.remove(file_path)
+if '--Clean' in argv:
+    print("======Clean======")
+    for file_name in os.listdir(zip_path):
+        file_path = os.path.join(zip_path, file_name)
+        os.remove(file_path)
+        print(file_path)
 
-print("======Build Engine======")
-os.chdir(engine_path)
-AsyncRun([engine_path + r"\_build.bat"])
+if '--BuildEngine' in argv:
+    print("======Build Engine======")
+    os.chdir(engine_path)
+    AsyncRun([engine_path + r"\_build.bat"])
 
-print("======Clean Engine======")
-AsyncRun([engine_path + r"\_clean.bat"])
+if '--CleanEngine' in argv:
+    print("======Clean Engine======")
+    AsyncRun([engine_path + r"\_clean.bat"])
 
-print("======Zip Engine======")
-args = [bandizip_path, "a", "-l:9", "-y", "-v:2000MB", "-t:60", engine_target_path, 
-        engine_path + r"\LocalBuilds\Engine",
-        ]
-AsyncRun(args)
+if '--ZipEngine' in argv:
+    print("======Zip Engine======")
+    args = [bandizip_path, "a", "-l:9", "-y", "-v:2000MB", "-t:60", engine_target_path, engine_path + r"\LocalBuilds\Engine",]
+    AsyncRun(args)
 
-print("======Zip Project======")
-args = [bandizip_path, "a", "-l:9", "-y", "-v:2000MB", "-t:60", project_target_path, 
-        project_path + r"\Art",
-        project_path + r"\Content",
-        project_path + r"\MooaToon_Project.uproject",
-        ]
-AsyncRun(args)
+if '--ZipProject' in argv:
+    print("======Zip Project======")
+    args = [bandizip_path, "a", "-l:9", "-y", "-v:2000MB", "-t:60", project_target_path,
+            project_path + r"\Art",
+            project_path + r"\Content",
+            project_path + r"\MooaToon_Project.uproject",]
+    AsyncRun(args)
 
 
-print("======Release======")
+current_date = datetime.date.today().strftime("%Y.%m.%d")
 file_paths = []
 for file_name in os.listdir(zip_path):
     file_path = os.path.join(zip_path, file_name)
     file_paths.append(file_path)
 
-current_date = datetime.date.today().strftime("%Y.%m.%d")
+if '--Release' in argv:
+    print("======Release======")
+    ghr.gh_release_create(
+        repo_name,
+        current_date,
+        publish=True,
+        name=current_date,
+        asset_pattern=file_paths,
+    )
 
-ghr.gh_release_create(
-    repo_name,
-    current_date,
-    publish=True,
-    name=current_date,
-    asset_pattern=file_paths,
-)
+if '--Reupload' in argv:
+    print("======Reupload======")
+    release_info = ghr.get_release_info(repo_name, current_date)
+
+    # 仅上传失败的文件
+    for asset in release_info['assets']:
+        if any(file.endswith(asset['name']) for file in file_paths):
+            file_paths.remove(asset['name'])
+
+    ghr.gh_asset_upload(repo_name, current_date, file_paths)
 
 
 print("Press any key to continue...")
